float to double

+++ app/src/main/java/pocket/pay/tp3_hci/model/Balance.kt
@@ -1,13 +1,9 @@
 package pocket.pay.tp3_hci.model
 
 import pocket.pay.tp3_hci.network.model.NetworkBalance
-import pocket.pay.tp3_hci.network.model.NetworkPayment
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
 
 data class Balance(
-    var balance: Float
+    var balance: Double
 ) {
     fun asNetworkModel(): NetworkBalance {
         return NetworkBalance(
